// src/app/Modules/rrhh/model/relplazpers.dto.ts
export interface RelPlazPersDTO {
    activoHoy:             string;
    adscripcionPzaCod:     string;
    adscripcionPzaDes:     string;
    aextinguir:            string;
    amortizar:             string;
    anioSalidaOpe:         number;
    apellido1:             string;
    apellido2:             string;
    apellidosNombre:       string;
    categoriaCod:          string;
    categoriaDes:          string;
    claseCod:              string;
    claseDes:              string;
    codPre:                number;
    codiTipoPlaza:         number;
    codigoRpt:             string;
    descripcionPza:        string;
    descripcionPzaDes:     string;
    dotacionVincplazaNum:  number;
    empleadoCod:           number;
    entidadDes:            string;
    escalaCod:             string;
    escalaDes:             string;
    expAdscFechaPza:       string;
    expAdscPza:            string;
    expFinPza:             string;
    expNombrFechaPza:      string;
    expNombrPza:           string;
    expeFinFechaPza:       Date;
    fechaCreaDotPlaza:     Date;
    finPosesionPza:        Date;
    finVersionFase:        Date;
    grupoCod:              string;
    grupoDes:              string;
    id:                    RelPlazPersIdDTO;
    inicioPosesionPza:     Date;
    inicioVersionFase:     Date;
    inicioVersionPlaza:    Date;
    inicioVersionPlazaDot: Date;
    jornadaReducida:       string;
    motivoAdscPza:         string;
    motivoCesePza:         string;
    motivoNombrPza:        string;
    naturalezaPzaDes:      string;
    naturalezaRelacionCod: string;
    naturalezaRelacionDes: string;
    nombre:                string;
    numPlazas:             number;
    obseFase:              string;
    organizacionNivel1Cod: string;
    organizacionNivel1Des: string;
    organizacionNivel2Cod: string;
    organizacionNivel2Des: string;
    organizacionNivel3Cod: string;
    organizacionNivel3Des: string;
    organizacionNivel4Cod: string;
    organizacionNivel4Des: string;
    organizacionNivel5Cod: string;
    organizacionNivel5Des: string;
    plazaAmortizada:       string;
    porcentajeJorn:        number;
    procesoOpe:            string;
    puestoVincplazaCod:    number;
    puestoVincplazaDes:    string;
    rptVincplazaCod:       string;
    rptVincplazaDes:       string;
    situacionAdminCod:     number;
    situacionAdminDes:     string;
    situacionPzaCod:       string;
    situacionPzaDes:       string;
    sometidaFunc:          string;
    subescalaCod:          string;
    subescalaDes:          string;
    tipoJornada:           string;
    versionDot:            number;
    versionPlaza:          number;
}

export interface RelPlazPersIdDTO {
    entidadCod:       number;
    faseNumero:       number;
    naturalezaPzaCod: string;
    numPlaza:         number;
    tipoPza:          number;
    versionFasePlaza: number;
}